feat: adiciona validações da API e ajustes finais

+++ src/utils/salvarPokemon.ts
@@ -1,33 +1,46 @@
-import inquirer from "inquirer";
 import { writeFile } from "fs/promises";
+
+import inquirer from "inquirer";
+
 import { Pokemon } from "../models/pokemon";
 
-async function salvarPokemon(
-  pokemon: Pokemon,
-  pokemonsNaPokedex: Pokemon[],
-) {
-  const desejaSalvar = await inquirer.prompt([
-    {
-      type: "select",
-      name: "salvar",
-      message: "Deseja salvar o pokémon na pokédex?",
-      choices: ["Sim", "Não"],
-    },
-  ]);
+async function salvarPokemon(pokemon: Pokemon, pokemonsNaPokedex: Pokemon[]) {
+  try {
+    const desejaSalvar = await inquirer.prompt([
+      {
+        type: "select",
+        name: "salvar",
+        message: "Deseja salvar o pokémon na pokédex?",
+        choices: ["Sim", "Não"],
+      },
+    ]);
 
-  if (desejaSalvar.salvar === "Sim") {
-    pokemonsNaPokedex.push(pokemon);
-    await writeFile("./pokedex.json", JSON.stringify(pokemonsNaPokedex), {
-      encoding: "utf-8",
-    });
+    if (desejaSalvar.salvar === "Sim") {
+      const pokemonExistePokedex = pokemonsNaPokedex.find(
+        (pokemonPokedex) => pokemon.nome === pokemonPokedex.nome,
+      );
 
-    console.log(`[OK] ${pokemon.name} salvo com sucesso!`);
-    return true;
-  }
+      if (pokemonExistePokedex) {
+        console.log(`[AVISO] ${pokemon.nome} já existe na Pokédex`);
+        return true;
+      }
 
-  if (desejaSalvar.salvar === "Não") {
-    console.log(`[AVISO] ${pokemon.name} não foi salvo na pokédex`);
-    return true;
+      pokemonsNaPokedex.push(pokemon);
+      await writeFile("./pokedex.json", JSON.stringify(pokemonsNaPokedex), {
+        encoding: "utf-8",
+      });
+
+      console.log(`[OK] ${pokemon.nome} salvo com sucesso!`);
+      return true;
+    }
+
+    if (desejaSalvar.salvar === "Não") {
+      console.log(`[AVISO] ${pokemon.nome} não foi salvo na pokédex`);
+      return true;
+    }
+  } catch {
+    console.error("[ERRO] Erro ao salvar o Pokémon");
+    return null;
   }
 }
 export { salvarPokemon };
